support for new reference types added

+++ opsrest/delete.py
@@ -16,9 +16,10 @@ def delete_resource(resource, schema, txn, idl):
 
     if resource.relation == OVSDB_SCHEMA_CHILD:
 
-        # get the row, delete its reference from the parent, delete the row
-        row = utils.get_row(resource.next, idl)
-        utils.delete_reference(row, resource, None, idl)
+        if resource.next.row is None:
+            raise Exception({'status' : httplib.METHOD_NOT_ALLOWED})
+
+        row = utils.delete_reference(resource.next, resource, schema, idl)
         row.delete()
 
     elif resource.relation == OVSDB_SCHEMA_REFERENCE:
